More header-file rearrangements
