+++ link-grammar/tokenize/tokenize.c
@@ -871,6 +871,21 @@ static void tokenization_done(Dictionary dict, Gword *altp)
 	}
 }
 
+static const char ** resize_alts(const char **arr, size_t len)
+{
+	arr = realloc(arr, (len+2) * sizeof(char *));
+	arr[len+1] = NULL;
+	return arr;
+}
+
+void altappend(Sentence sent, const char ***altp, const char *w)
+{
+	size_t n = altlen(*altp);
+
+	*altp = resize_alts(*altp, n);
+	(*altp)[n] = string_set_add(w, sent->string_set);
+}
+
 /*
 	Here's a summary of how subscripts are handled:
