+++ link-grammar/utilities.h
@@ -453,7 +453,6 @@ void dyn_str_delete(dyn_str*);
 void dyn_strcat(dyn_str*, const char*);
 
 size_t altlen(const char **);
-void altappend(Sentence, const char ***, const char *);
 
 /* routines for allocating basic objects */
 void init_memusage(void);
